:art: Rename Register to RegistryService
- Keep but deprecate Register class for compatibility
